requirement.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==3.2.12
djangorestframework==3.13.1
idna==3.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
protobuf==3.20.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
pycoingecko==2.2.0
gunicorn==20.1.0
websocket-client==1.3.3
pycryptodome~=3.15.0
